Update GenerateServiceProvider.php
修复少判断bug

+++ src/GenerateServiceProvider.php
@@ -23,10 +23,10 @@ class GenerateServiceProvider extends ServiceProvider
     public function boot()
     {
         if (false == is_file(public_path('lock.php'))) {
-            is_file(database_path('migrations/2014_10_12_000000_create_users_table.php')){
+            if (is_file(database_path('migrations/2014_10_12_000000_create_users_table.php'))) {
                 unlink(database_path('migrations/2014_10_12_000000_create_users_table.php'));
             }
-            is_file(database_path('migrations/2014_10_12_100000_create_password_resets_table.php')){
+            if (is_file(database_path('migrations/2014_10_12_100000_create_password_resets_table.php'))){
                 unlink(database_path('migrations/2014_10_12_100000_create_password_resets_table.php'));
             }
             $this->publishes([
